# internal/serve/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "jevql serve",
    "version": "1",
    "description": "The HTTP API of `jevql serve`: run jev-enabled SQL against Postgres, judge rows you already hold, inspect the schema and the answer cache. Errors always return an ErrorBody whose `code` is one of `sql` (400/404), `budget` (402), `auth` (401/403), `api` (502) or `internal` (500)."
  },
  "servers": [{ "url": "http://127.0.0.1:7433" }],
  "security": [{ "bearer": [] }],
  "paths": {
    "/v1/health": {
      "get": {
        "summary": "Liveness and version",
        "operationId": "health",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Health" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/query": {
      "post": {
        "summary": "Run one SQL statement (plain or with jev_* calls)",
        "operationId": "query",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/QueryRequest" } } } },
        "responses": {
          "200": { "description": "Result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/QueryResult" } } } },
          "400": { "$ref": "#/components/responses/SQLError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "402": { "$ref": "#/components/responses/Budget" },
          "502": { "$ref": "#/components/responses/APIError" }
        }
      }
    },
    "/v1/explain": {
      "post": {
        "summary": "Plan and cost estimate for a jev statement, without calling TypeSafe",
        "operationId": "explain",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/QueryRequest" } } } },
        "responses": {
          "200": { "description": "Explain", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Explain" } } } },
          "400": { "$ref": "#/components/responses/SQLError" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/judge": {
      "post": {
        "summary": "Judge rows you already hold, with no database involved",
        "description": "Shares batching, dedup and the answer cache with the SQL path. Answers come back in input order.",
        "operationId": "judge",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JudgeRequest" } } } },
        "responses": {
          "200": { "description": "Answers", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JudgeResult" } } } },
          "400": { "$ref": "#/components/responses/SQLError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "402": { "$ref": "#/components/responses/Budget" },
          "502": { "$ref": "#/components/responses/APIError" }
        }
      }
    },
    "/v1/cache": {
      "get": {
        "summary": "Answer cache statistics",
        "operationId": "cacheInfo",
        "responses": {
          "200": { "description": "Cache info", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CacheInfo" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "delete": {
        "summary": "Clear the answer cache",
        "operationId": "cacheClear",
        "responses": {
          "200": { "description": "Cleared", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CacheCleared" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/v1/schema/tables": {
      "get": {
        "summary": "List user relations",
        "operationId": "listTables",
        "responses": {
          "200": { "description": "Relations", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/TableInfo" } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/schema/tables/{name}": {
      "get": {
        "summary": "Describe one relation (columns and indexes)",
        "operationId": "describeTable",
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Relation name, optionally schema-qualified" }],
        "responses": {
          "200": { "description": "Relation detail", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TableDetail" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "No such relation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorBody" } } } }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "This document",
        "operationId": "openapi",
        "security": [],
        "responses": { "200": { "description": "OpenAPI 3.1 document", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    }
  },
  "components": {
    "securitySchemes": { "bearer": { "type": "http", "scheme": "bearer", "description": "Required when the server was started with --token or JEVQL_TOKEN." } },
    "responses": {
      "Unauthorized": { "description": "Missing or invalid bearer token (code auth)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorBody" } } } },
      "Forbidden": { "description": "Feature disabled on this server (code auth)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorBody" } } } },
      "SQLError": { "description": "Parse, rewrite, validation or Postgres error (code sql)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorBody" } } } },
      "Budget": { "description": "--max-rows or --max-chars fired before any TypeSafe call (code budget)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorBody" } } } },
      "APIError": { "description": "TypeSafe returned an error (code api)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorBody" } } } }
    },
    "schemas": {
      "Health": {
        "type": "object",
        "properties": { "ok": { "type": "boolean" }, "version": { "type": "string" }, "model": { "type": "string" } },
        "required": ["ok", "version", "model"]
      },
      "ErrorBody": {
        "type": "object",
        "properties": { "error": { "type": "string" }, "code": { "type": "string", "enum": ["sql", "budget", "api", "auth", "internal"] } },
        "required": ["error", "code"]
      },
      "QueryRequest": {
        "type": "object",
        "properties": {
          "sql": { "type": "string" },
          "threshold": { "type": "number", "description": "Overrides the default jev() threshold for this request" },
          "explain": { "type": "boolean", "description": "Return the plan only; no TypeSafe calls" },
          "max_rows": { "type": "integer", "description": "Overrides --max-rows for this request" }
        },
        "required": ["sql"]
      },
      "Stats": {
        "type": "object",
        "properties": {
          "collect_rows": { "type": "integer" }, "judged": { "type": "integer" }, "requests": { "type": "integer" },
          "cache_hits": { "type": "integer" }, "input_tokens": { "type": "integer" }, "output_tokens": { "type": "integer" },
          "usd": { "type": "number" }, "elapsed_ms": { "type": "integer" }
        },
        "required": ["collect_rows", "judged", "requests", "cache_hits", "input_tokens", "output_tokens", "usd", "elapsed_ms"]
      },
      "Explain": {
        "type": "object",
        "properties": {
          "collect_sql": { "type": "string" }, "rows": { "type": "integer" }, "sources": { "type": "integer" },
          "questions": { "type": "integer" }, "judgements": { "type": "integer" }, "batches": { "type": "integer" },
          "avg_row_chars": { "type": "number" }, "tokens": { "type": "integer" }, "usd": { "type": "number" },
          "server_order": { "type": "boolean" }, "question_list": { "type": "array", "items": { "type": "string" } }
        },
        "required": ["collect_sql", "rows", "sources", "questions", "judgements", "batches", "avg_row_chars", "tokens", "usd", "server_order", "question_list"]
      },
      "QueryResult": {
        "type": "object",
        "properties": {
          "columns": { "type": "array", "items": { "type": "string" } },
          "rows": { "type": "array", "items": { "type": "array", "items": {} } },
          "row_count": { "type": "integer" },
          "tag": { "type": "string", "description": "Postgres command tag, e.g. SELECT 3 or INSERT 0 1" },
          "jev": { "type": "boolean", "description": "Whether jev_* functions were evaluated" },
          "stats": { "oneOf": [{ "$ref": "#/components/schemas/Stats" }, { "type": "null" }] },
          "explain": { "oneOf": [{ "$ref": "#/components/schemas/Explain" }, { "type": "null" }] }
        },
        "required": ["columns", "rows", "row_count", "tag", "jev", "stats", "explain"]
      },
      "JudgeRequest": {
        "type": "object",
        "properties": {
          "question": { "type": "string" },
          "kind": { "type": "string", "enum": ["noul", "choice", "score"], "description": "Defaults to noul" },
          "options": { "type": "array", "items": { "type": "string" }, "description": "Choice options or score levels" },
          "threshold": { "type": "number", "description": "noul: pass when p >= threshold" },
          "rows": { "type": "array", "items": { "type": "object", "additionalProperties": {} } },
          "raw": { "type": "boolean", "description": "Include the raw TypeSafe answer in each JudgeAnswer" }
        },
        "required": ["question", "rows"]
      },
      "JudgeAnswer": {
        "type": "object",
        "properties": {
          "p": { "type": "number", "description": "noul probability" },
          "pass": { "type": "boolean", "description": "noul: p >= threshold" },
          "choice": { "type": "string" },
          "score": { "type": "number", "description": "score: weighted level index" },
          "norm": { "type": "number", "description": "score / (levels - 1)" },
          "confidence": { "type": "number" },
          "probabilities": { "type": "object", "additionalProperties": { "type": "number" } },
          "raw": { "type": "object", "additionalProperties": {} }
        }
      },
      "JudgeResult": {
        "type": "object",
        "properties": {
          "answers": { "type": "array", "items": { "$ref": "#/components/schemas/JudgeAnswer" } },
          "stats": { "oneOf": [{ "$ref": "#/components/schemas/Stats" }, { "type": "null" }] }
        },
        "required": ["answers", "stats"]
      },
      "CacheInfo": {
        "type": "object",
        "properties": { "entries": { "type": "integer" }, "path": { "type": "string" } },
        "required": ["entries"]
      },
      "CacheCleared": {
        "type": "object",
        "properties": { "deleted": { "type": "integer" } },
        "required": ["deleted"]
      },
      "TableInfo": {
        "type": "object",
        "properties": { "schema": { "type": "string" }, "name": { "type": "string" }, "kind": { "type": "string" } },
        "required": ["schema", "name", "kind"]
      },
      "ColumnInfo": {
        "type": "object",
        "properties": { "name": { "type": "string" }, "type": { "type": "string" }, "nullable": { "type": "boolean" }, "default": { "type": "string" } },
        "required": ["name", "type", "nullable", "default"]
      },
      "IndexInfo": {
        "type": "object",
        "properties": { "name": { "type": "string" }, "definition": { "type": "string" } },
        "required": ["name", "definition"]
      },
      "TableDetail": {
        "type": "object",
        "properties": {
          "schema": { "type": "string" }, "name": { "type": "string" }, "kind": { "type": "string" },
          "columns": { "type": "array", "items": { "$ref": "#/components/schemas/ColumnInfo" } },
          "indexes": { "type": "array", "items": { "$ref": "#/components/schemas/IndexInfo" } }
        },
        "required": ["schema", "name", "kind", "columns", "indexes"]
      }
    }
  }
}